the first 50 fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,18 +5,18 @@
  */
 int main(void)
 {
-	int prev = 0;
-	int curr = 1;
+	long int prev = 0;
+	long int curr = 1;
 	int count = 0;
-	int temp;
+	long int temp;
 
 	while (count < 50)
 	{
-		printf("%lu, ",prev + curr);
+		printf("%lu, ", prev + curr);
 		temp = curr;
 		curr = prev + curr;
 		prev = temp;
-		count ++;
+		count++;
 	}
 	return (0);
 }
